Bootstrapping and logging updates.
  - add a "project type" in order to define and bootstrap
    different types of projects. The "default" project
    type will be called if no other is specified.
  - check for the files to copy in pwd, home config dir,
    and install dir (in that order).
  - add a "-noreplace" flag for copy files. This will prevent
    the bootstrap process from modifying any of their contents,
    can also be used to copy binaries, for whatever reason.
  - update the copy files with the variables provided during
    the bootstrap process. e.g. if the variable in the config
    is "ssh_port" it will look for and replace "${SSH_PORT}
    in all of the files copied to the bootstrapped project
    (unless the file is tagged with "-noreplace").
  - write main log file to the user config dir
  - set up project logging via script. This will start a new
    log file for each session created. When exiting the project
    directory, the script session will exit. For now this
    defaults the pwd back to the root of the project directory,
    hopefully this is a temporary issue.

+++ iet/bootstrap.py
@@ -19,6 +19,7 @@ def parse_args(print_help=False):
     argparser = argparse.ArgumentParser(prog='bootstrap')
     argparser.add_argument('-c', '--config', help='Location of config file', default=None)
     argparser.add_argument('-b', '--basename', help='directory to bootstrap', default=os.getcwd())
+    argparser.add_argument('-pt', '--project_type', help='project directory name', default='default')
     argparser.add_argument('-pn', '--project_name', help='project directory name', default=None)
     args = argparser.parse_args()
     if print_help:
@@ -26,7 +27,7 @@ def parse_args(print_help=False):
     else:
         return args
 
-
+# I should refactor this to another module cause imma use it more than here
 def load_config(config, section=None):
     config = open(config, 'r')
     config = yaml.load(config)
@@ -43,9 +44,42 @@ def mkvars(vars):
     logger.debug("vars: %s" % vars)
     for var in vars:
         iet_vars[var] = text = input(Fore.MAGENTA + " [+] Please enter value for %s: " % var + Style.RESET_ALL)
+    return iet_vars
 
 
-def main(config, basename, project_name):
+def get_file_path(copy_file):
+    """
+    Find the first instance of the specified file and return the path.
+    """
+    if os.path.isfile(copy_file):
+        pass
+    elif os.path.isfile(os.path.join(Path.home(), '.iet', copy_file)):
+        copy_file = os.path.join(Path.home(), '.iet', copy_file)
+    elif pkg_resources.resource_filename(__name__, copy_file):
+        copy_file = pkg_resources.resource_filename(__name__, copy_file)
+    return copy_file
+
+
+def do_copy_update_file(copy_file, dst, iet_vars):
+    """
+    For the copy file, check iet_vars and replace the variables if
+    present in the copy file.
+    """
+    with open(copy_file, 'r') as copy_data:
+        copy_data = copy_data.read()
+    for k, v in iet_vars.items():
+        var = str("${%s}" % k.upper())
+        logger.debug("Searching for: %s" % var)
+        if var in copy_data:
+            logger.debug("Found %s in %s" % (var, copy_file))
+            copy_data = copy_data.replace(var, v)
+    with open(dst, 'w') as outfile:
+        logger.debug("Writing %s" % dst)
+        outfile.write(copy_data)
+    return True
+
+
+def main(config, basename, project_name, project_type):
     """
     - load config file
     - create base directory and subdirs from config.
@@ -56,9 +90,12 @@ def main(config, basename, project_name):
     """
     global logger
     logger = setup_logging(load_config(config, 'logging'))
-    config = load_config(config, 'bootstrap')
+    config = load_config(config, 'bootstrap')[project_type]
+    logger.info("project type: %s" % project_type)
     logger.info("basename: %s" % basename,)
     logger.info("project name: %s" % project_name)
+    iet_vars = mkvars(config['vars'])
+    logger.debug("iet variables: %s" % iet_vars)
     for _dir in config['dirs']:
         logger.debug("Creating directory: %s" % _dir)
         os.makedirs(os.path.join(basename, project_name, _dir))
@@ -70,24 +107,22 @@ def main(config, basename, project_name):
                 dst = os.path.join(basename, project_name, src)
             else:
                 dst = os.path.join(basename, project_name, file_dirs, src)
-            # TODO: this could be a cleaner check
-            try:
+            do_replace = True
+            if copy_file.endswith('-noreplace'):
+                do_replace = False
+                copy_file = copy_file.replace('-noreplace', '')
+            copy_file = get_file_path(copy_file)
+            if do_replace:
+                if do_copy_update_file(copy_file, dst, iet_vars):
+                    logger.debug("Copied %s to %s" % (copy_file, dst))
+                else:
+                    logger.error("ERROR: could not copy %s to %s" % (copy_file, dst))
+            else:
+                logger.debug("Not performing find/replace on %s" % copy_file)
                 copyfile(copy_file, dst)
                 logger.debug("Copied %s to %s" % (copy_file, dst))
-            except FileNotFoundError:
-                if os.path.isfile(os.path.join(Path.home(), '.iet', copy_file)):
-                    copy_file = os.path.join(Path.home(), '.iet', copy_file)
-                    copyfile(copy_file, dst)
-                    logger.debug("Copied %s to %s" % (copy_file, dst))
-                elif pkg_resources.resource_filename(__name__, copy_file):
-                    copy_file = pkg_resources.resource_filename(__name__, copy_file)
-                    copyfile(copy_file, dst)
-                    logger.debug("Copied %s to %s" % (copy_file, dst))
-    # I don't know if this should go before or after the file/folder creation
-    # maybe get the vars first, then open/write the .in/.out file instead of moving and sed'ing
-    mkvars(config['vars'])
 
 
 if __name__ == "__main__":
     args = parse_args()
-    main(args.config, args.basename, args.project_name)
+    main(args.config, args.basename, args.project_name, args.project_type)
